Move Suggestion API to top-level
Refactor the session bearer auth to a shared middleware
