+++ src/api/main.ts
@@ -2,10 +2,11 @@ import { Hono } from 'hono'
 import { cors } from 'hono/cors'
 import adminApp from './admin.ts'
 import inviteApp from './invite.ts'
-import loginApp from './login.ts'
-import userApp from './user.ts'
 import type { KvProvidedVariables } from './kv.ts'
+import loginApp from './login.ts'
 import { listenQueue } from './queue.ts'
+import suggestionApp from './suggestion.ts'
+import userApp from './user.ts'
 
 const kv = await Deno.openKv()
 kv.listenQueue((msg) => listenQueue(kv, msg))
@@ -22,6 +23,7 @@ const app = new Hono<{ Variables: KvProvidedVariables }>()
   .route('/admin', adminApp)
   .route('/invite', inviteApp)
   .route('/login', loginApp)
+  .route('/suggestion', suggestionApp)
   .route('/user', userApp)
 
 Deno.serve(app.fetch)
